requirements.txt
streamlit
faiss-cpu
pypdf
mistralai